# docs/openapi.yaml
openapi: 3.1.0
info:
  title: THCPN IoT Research Platform API
  version: 0.15.0
  description: |
    THCPN 科研物联网数据管理平台后端 API。

    当前文档覆盖已实现的阶段功能：健康检查、短信验证码登录/注册、邮箱验证码验证、密码注册/登录、TOTP MFA、refresh token、logout、会话管理、
    JWT 认证、Workspace 列表/创建、Workspace 成员管理、Project、Site、Device、DeviceRelation、DataStream、DataSource、DataStreamBinding、Telemetry Query、Media Query、Dataset 元信息与遥测查询管理和 ExportJob 导出任务、
    AccessGrant 外部分享/售后临时授权、Invitation、AuditLog、Camera 实时预览，以及系统管理员后台 DataSource / THCPN 标准站和组网站拓扑同步接口。

    正式业务接口优先使用 `Authorization: Bearer <token>`。
    开发环境可通过配置启用 `X-User-ID` fallback；生产环境应关闭该能力。
servers:
  - url: http://127.0.0.1:8080
    description: Local development API server
...
paths:
...
  /api/v1/me:
    get:
      tags:
        - Auth
      operationId: getMe
      summary: Get current authenticated user
      security:
        - bearerAuth: []
        - devUserHeader: []
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
              examples:
                default:
                  value:
                    user:
                      id: 018fa0ec-9b6e-7a3d-9c7b-1d8b04f31988
                      name: 张三
                      phone: "13800000000"
                      email: zhangsan@example.com
                      status: active
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [Auth]
      operationId: updateMe
      summary: Update the current user's profile
      security:
        - bearerAuth: []
        - devUserHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: Updated current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/public/devices/{public_slug}/unlock:
    post:
      tags: [Devices]
      operationId: unlockPublicDevice
      summary: Unlock a protected public device for seven days
      security: []
      parameters: [{ name: public_slug, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties: { password: { type: string, minLength: 8, maxLength: 72 } }
      responses:
        "204": { description: Device unlocked in this browser. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/devices/{public_slug}/telemetry:
    get:
      tags: [Telemetry]
      operationId: queryPublicDeviceTelemetry
      summary: Query the latest 72 hours of public telemetry
      security: []
      parameters: [{ name: public_slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Adaptively sampled public telemetry.
          content: { application/json: { schema: { $ref: "#/components/schemas/TelemetryQueryResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/public/devices/{public_slug}/images:
    get:
      tags: [Media]
      operationId: listPublicDeviceImages
      summary: List preview-only images from the latest 72 hours
      security: []
      parameters:
        - { name: public_slug, in: path, required: true, schema: { type: string } }
        - { name: data_stream_id, in: query, required: true, schema: { $ref: "#/components/schemas/UUID" } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 24 } }
      responses:
        "200":
          description: Public image previews.
          content: { application/json: { schema: { $ref: "#/components/schemas/MediaListResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token returned by SMS or password login APIs.
    devUserHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: Development-only fallback authentication. Production should disable it.
...
  responses:
    InvalidArgument:
      description: Invalid request argument or body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            invalid:
              value:
                error:
                  code: INVALID_ARGUMENT
                  message: invalid request body
                  request_id: req_123
    Unauthorized:
      description: Missing, invalid or inactive authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            unauthorized:
              value:
                error:
                  code: UNAUTHORIZED
                  message: invalid bearer token
                  request_id: req_123
...
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            notFound:
              value:
                error:
                  code: NOT_FOUND
                  message: workspace member not found
                  request_id: req_123
...
    RateLimited:
      description: Request was rate-limited or the account is temporarily locked.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            limited:
              value:
                error:
                  code: RATE_LIMITED
                  message: sms code requested too frequently
                  request_id: req_123
...
    Internal:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            internal:
              value:
                error:
                  code: INTERNAL
                  message: internal error
                  request_id: req_123
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required:
        - code
        - message
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        request_id:
          type: string
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - UNAUTHORIZED
        - PERMISSION_DENIED
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - DATA_SOURCE_ERROR
        - INTERNAL
        - SERVICE_UNAVAILABLE
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Phone:
      type: string
      description: Normalized phone number. Lightweight validation allows digits and an optional leading plus sign.
      examples:
        - "13800000000"
        - "+8613800000000"
    Email:
      type: string
      format: email
    UserStatus:
      type: string
      enum:
        - active
        - disabled
...
    MediaType:
      type: string
      enum:
        - image
        - video
        - audio
...
    Actor:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        phone:
          $ref: "#/components/schemas/Phone"
        email:
          $ref: "#/components/schemas/Email"
        status:
          $ref: "#/components/schemas/UserStatus"
        phone_verified_at:
          $ref: "#/components/schemas/Timestamp"
        email_verified_at:
          $ref: "#/components/schemas/Timestamp"
    MeResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/Actor"
...
    TelemetryPoint:
      type: object
      required:
        - ts
        - value
        - quality
      properties:
        ts:
          $ref: "#/components/schemas/Timestamp"
        value:
          type: number
          format: double
        quality:
          type: string
    QueryWarning:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        count:
          type: integer
          format: int32
    TelemetrySeries:
      type: object
      required:
        - data_stream_id
        - code
        - name
        - points
        - source_count
        - returned_count
        - sampled
        - complete
      properties:
        data_stream_id:
          $ref: "#/components/schemas/UUID"
        code:
          type: string
        name:
          type: string
        unit:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/TelemetryPoint"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/QueryWarning"
        source_count:
          type: integer
          format: int32
          description: Number of valid source points found in the requested time range when known.
        returned_count:
          type: integer
          format: int32
          description: Number of points included in this response.
        sampled:
          type: boolean
          description: Whether the returned points were reduced for chart rendering.
        complete:
          type: boolean
          description: Whether the full requested time range was scanned.
    TelemetryQueryResponse:
      type: object
      required:
        - device_id
        - start_time
        - end_time
        - limit
        - series
      properties:
        device_id:
          $ref: "#/components/schemas/UUID"
        start_time:
          $ref: "#/components/schemas/Timestamp"
        end_time:
          $ref: "#/components/schemas/Timestamp"
        limit:
          type: integer
          format: int32
        series:
          type: array
          items:
            $ref: "#/components/schemas/TelemetrySeries"
    MediaItem:
      type: object
      required:
        - id
        - device_id
        - data_stream_id
        - captured_at
        - media_type
        - preview_url
        - download_allowed
        - delete_allowed
      properties:
        id:
          type: string
        device_id:
          $ref: "#/components/schemas/UUID"
        data_stream_id:
          $ref: "#/components/schemas/UUID"
        captured_at:
          $ref: "#/components/schemas/Timestamp"
        media_type:
          $ref: "#/components/schemas/MediaType"
        thumbnail_url:
          type: string
        preview_url:
          type: string
        download_allowed:
          type: boolean
        download_url:
          type: string
        delete_allowed:
          type: boolean
        delete_url:
          type: string
    MediaListResponse:
      type: object
      required:
        - items
        - page
        - page_size
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        total:
          type: integer
          format: int32
...